+++ src/toast/ToastItemContext.tsx
@@ -1,4 +1,4 @@
-import {createContext} from "react";
+import {createContext, useContext} from "react";
 
 type ToastItemContextValue = {
   toastId: string;
@@ -10,4 +10,14 @@ const ToastItemContext = createContext<ToastItemContextValue>({
 
 ToastItemContext.displayName = "ToastItemContext";
 
-export {ToastItemContext};
+function useToastItemContext() {
+  const toastItemContext = useContext(ToastItemContext);
+
+  if (!toastItemContext) {
+    throw new Error("Trying to consume ToastItemContext outside of its provider");
+  }
+
+  return toastItemContext;
+}
+
+export {ToastItemContext, useToastItemContext};
