+++ src/toast/util/toastUtils.ts
@@ -1,4 +1,4 @@
-import {ToastData} from "./toastTypes";
+import {ToastData} from "./context/toastContextTypes";
 
 function isSameToast(toastId: string) {
   return (toast2: ToastData) => toastId === toast2.id;
